# Conflicts:
#	include/Empire/level.h
#	src/entities.c
#	src/level.c

+++ include/Empire/level.h
@@ -10,7 +10,7 @@ struct emp_asset_t;
 typedef struct emp_value_buffer_t
 {
 	u8 values[emp_value_buffer_capacity];
-	size_t count;
+	u32 count;
 } emp_value_buffer_t;
 
 typedef struct emp_tile_desc_t
@@ -25,14 +25,14 @@ typedef struct emp_value_grid_t
 	u32 grid_width;
 	u32 grid_height;
 	u8* entries;
-	size_t count;
+	u32 count;
 } emp_value_grid_t;
 
 typedef struct emp_tiles_data_t
 {
 	char* tilemap;
 	emp_tile_desc_t* values;
-	size_t count;
+	u32 count;
 } emp_tiles_data_t;
 
 typedef struct emp_sublevel_t
@@ -46,13 +46,13 @@ typedef struct emp_sublevel_t
 typedef struct emp_sublevel_list_t
 {
 	emp_sublevel_t entries[emp_sublevel_capacity];
-	size_t count;
+	u32 count;
 } emp_sublevel_list_t;
 
 typedef struct emp_tile_desc_list_t
 {
 	emp_tile_desc_t entries[emp_value_buffer_capacity];
-	size_t count;
+	u32 count;
 } emp_tile_desc_list_t;
 
 typedef enum emp_entity_type {
@@ -114,7 +114,7 @@ typedef struct emp_level_asset_t
 } emp_level_asset_t;
 
 u32 emp_level_query(emp_level_asset_t* level, emp_entity_type type, u32 offset);
-emp_level_entity_t* emp_level_get(emp_level_asset_t* level, size_t at);
+emp_level_entity_t* emp_level_get(emp_level_asset_t* level, u32 at);
 
 u32 emp_level_teleporter_list_find(emp_level_teleporter_list_t* set, u64 id);
